pyproject.toml
[project]
name = "infraguard"
version = "0.1.0"
description = "Red team infrastructure tracker and C2 redirector"
readme = "README.md"
requires-python = ">=3.12"
license = "BSD-2-Clause"
dependencies = [
    "click>=8.1",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "httpx>=0.27",
    "starlette>=0.36",
    "uvicorn[standard]>=0.27",
    "aiosqlite>=0.20",
    "cryptography>=42.0",
]

[project.optional-dependencies]
web = ["fastapi>=0.109", "jinja2>=3.1"]
tui = ["textual>=0.47"]
geoip = ["maxminddb>=2.5"]
dns = ["aiodns>=3.1"]
all = ["infraguard[web,tui,geoip,dns]"]

[project.scripts]
infraguard = "infraguard.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["infraguard"]
